# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Doctor Work Platform API
  version: 1.0.0
  description: |
    Doctor workspace for the hospital: authentication, patient records, EMR with
    order validation, consultations, remote meetings, health management, and audit.

    ## Conventions

    - All business routes live under `/api`.
    - Success: `{ "code": 0, "message": "ok", "data": ... }` — HTTP 200 with `code`
      pinned to `0`.
    - Failure: `{ "code": <http status>, "message": "...", "data": null }` — on a
      failure, and only on a failure, `code` equals the HTTP status code.
    - Paginated lists take `page` (default 1) and `size` (default 20, max 100) and
      return `{ items, total, page, size }` inside `data`.
    - A list operation that declares **no** `page`/`size` parameters returns its rows
      as a flat array in `data` instead. These are the fixed-size dictionaries
      (`/roles`, `/departments`, `/allergens`, `/drugs`, `/emr/templates`) and the
      per-parent collections bounded by one parent (`/emr/orders`,
      `/emr/records/{id}/versions`, `/patient-groups`,
      `/patients/{patient_no}/histories`, `/patients/{patient_no}/allergies`,
      `/patients/{patient_no}/assessments`, `/meetings/{id}/participants`,
      `/meetings/{id}/materials`). Read the operation's `data` schema rather than
      assuming the paginated shape.
    - `message` is an English developer-facing string. Clients must render their
...
    - Timestamps are ISO 8601 in UTC.
...

...

...

...

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/departments:
    get:
      operationId: listDepartments
      tags: [Users]
      summary: List departments
      description: |
        Already implemented, and the one endpoint in this contract that **carries no task
        ID**: it predates the task register in `project_plan.md` and no sign-off standard
        asserts anything about it, so attributing it to one would be inventing an owner.
        `API-索引.md` marks the same row "已有" for the same reason. It is listed here
        because M03 §4 asks every interface to be traceable, and the honest trace is this
        note rather than a made-up T-number.

        It feeds the department pickers and filters that T09 (department isolation) and
        T13 (patient search) depend on, so it must stay available to any authenticated
        caller — `junior` included.
      responses:
        '200':
          description: Departments.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentListResponse' }

        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/patients:
    get:
      operationId: listPatients
      tags: [Patients]
      summary: Search patients
      description: |
        Task T13. The four search conditions — fuzzy name, exact patient number,
        symptom tag, and admission-date range — combine freely with AND, and all four
        may be sent at once. With none set, the full page for the caller's department is
        returned. Results default to `created_at` descending.

        A non-admin caller is silently restricted to their own department (T09); passing
        another department does not widen the result set.
      parameters:
        - name: name
          in: query
          description: Fuzzy match on the patient name.
          schema: { type: string }
        - name: patient_no
          in: query
          description: Exact match. Distinct from `name`, which is fuzzy.
          schema: { type: string, example: P20260001 }
        - name: symptom_tags
          in: query
          description: |
            Matches `symptom_tags` on the patient. Repeat the parameter to match any
            of several tags; the name is plural for that reason, and a client sending
            the singular `symptom_tag` gets no filtering at all rather than an error.

            This is one of four search conditions — `name`, `patient_no`,
            `symptom_tags`, and the admission-date range. They combine with AND and
            may be used in any combination; all four empty returns the full page.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
            example: [胸痛, 高血压]
        - name: department
          in: query
          schema: { type: string }
        - name: group_id
          in: query
          description: |
            Filter to one group. This **stacks with** the other conditions rather than
            replacing them: T17 scenario S1 filters to a two-member group and then adds
            a name search, and expects one row, not two. Treating the two as mutually
            exclusive fails sign-off.
          schema: { type: integer }
        - name: admitted_from
          in: query
          description: Start of the admission-date range, inclusive.
          schema: { type: string, format: date }
        - name: admitted_to
          in: query
          description: End of the admission-date range, inclusive.
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of patients. An empty result is an empty `items` array, not a 404.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
    post:
      operationId: createPatient
      tags: [Patients]
      summary: Create a patient
      description: |
        Task T13. The server generates and returns the unique `patient_no`; the client
        does not supply one. A missing required field is a 422 whose `message` names the
        field, per T13 scenario S2.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientCreateRequest' }
      responses:
        '200':
          description: Created, with the generated `patient_no` in the body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/patients/{patient_no}:
    parameters:
      - $ref: '#/components/parameters/PatientNo'
    get:
      operationId: getPatient
      tags: [Patients]
      summary: Read a patient with allergies, history, and groups
      description: |
        Task T14. Subject to department scope. Returns 404 rather than 403 when the
        patient belongs to another department, so the status code cannot be used to
        probe for records.
      responses:
        '200':
          description: The patient.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePatient
      tags: [Patients]
      summary: Edit patient basics
      description: |
        Task T13. Editing patient information is an explicit instructor requirement
        and has no endpoint in the current backend.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatientUpdateRequest' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      operationId: deletePatient
      tags: [Patients]
      summary: Delete a patient
      description: |
        Task T13. The current backend exposes this without authentication; it must
        require a signed-in user before any demo on a shared network.

        **Deletion strategy (T13 criterion 1 requires this to be written down):** a
        soft delete. The row is flagged as removed and disappears from list and detail
        reads, while the audit trail and any archived record that references the patient
        stay intact. A hard `DELETE` would break the append-only audit guarantee and the
        archive lock, so it is not used. Child rows — allergies, histories, group
        memberships — are left in place and filtered out with the parent.
      responses:
        '200':
          description: Soft-deleted; the patient is no longer readable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkData' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/patients/{patient_no}/allergies:
    parameters:
      - $ref: '#/components/parameters/PatientNo'
    get:
      operationId: listPatientAllergies
      tags: [Patients]
      summary: List a patient's allergies
      description: |
        Task T14. The patient detail page keeps a persistent red warning banner for
        severe allergies rather than hiding them in a tab, and shows no banner at all
        when the list is empty.

        This same list is what the order-validation engine reads, through one shared
        function `get_patient_allergens(patient_no)`. The banner and the blocking engine
        must not query allergies separately — T14 criterion 2 makes that a sign-off
        condition, because two sources drift.
      responses:
        '200':
          description: Allergies.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllergyListResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createPatientAllergy
      tags: [Patients]
      summary: Record an allergy
      description: |
        Task T14. Adding, changing, and deleting an allergy each write an audit entry
        (`allergy.create` / `allergy.update` / `allergy.delete`); the delete entry names
        the allergen that was removed, which is what T14 scenario S2 reads back.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AllergyCreateRequest' }
      responses:
        '200':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllergyResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /api/allergens:
    get:
      operationId: listAllergens
      tags: [Patients]
      summary: List the allergen dictionary
      description: |
        Task T14. Feeds the allergen picker so spellings stay consistent with the drug
        contraindications that the order-validation engine compares against.

        The dictionary must cover at least penicillins, sulfonamides, cephalosporins,
        aspirin, and contrast media, and must allow a free-text addition for anything
        outside it.
      responses:
        '200':
          description: Allergens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllergenDictionaryResponse' }

        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token from `POST /api/auth/verify-code`. Valid for 2 hours.
        Logout blacklists the token's `jti` in Redis so it stops working immediately.
...